refactor(ledger): separate ledger hash mod
