+++ rust/src/ledger/mod.rs
@@ -2,6 +2,7 @@ pub mod account;
 pub mod coinbase;
 pub mod diff;
 pub mod genesis;
+pub mod hash;
 pub mod public_key;
 pub mod staking;
 pub mod store;
@@ -15,16 +16,14 @@ use crate::{
         diff::LedgerDiff,
         public_key::PublicKey,
     },
-    protocol::serialization_types::{
-        common::{Base58EncodableVersionedType, HashV1},
-        version_bytes,
-    },
 };
-use anyhow::bail;
 use diff::account::AccountDiff;
 use log::debug;
 use serde::{Deserialize, Serialize};
 use std::{collections::HashMap, str::FromStr};
+
+// re-export [hash::LedgerHash]
+pub type LedgerHash = hash::LedgerHash;
 
 #[derive(Default, Clone, Serialize, Deserialize)]
 pub struct Ledger {
@@ -48,56 +47,6 @@ impl Ledger {
 #[derive(Default, Clone, Serialize, Deserialize)]
 pub struct NonGenesisLedger {
     pub ledger: Ledger,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct LedgerHash(pub String);
-
-impl LedgerHash {
-    pub const PREFIX: &'static [&'static str] = &["jx", "jw", "jy", "jz"];
-    pub const LEN: usize = 51;
-
-    pub fn from_hashv1(hashv1: HashV1) -> Self {
-        let versioned: Base58EncodableVersionedType<{ version_bytes::LEDGER_HASH }, _> =
-            hashv1.into();
-        Self(versioned.to_base58_string().unwrap())
-    }
-
-    pub fn from_bytes(bytes: Vec<u8>) -> anyhow::Result<Self> {
-        let hash = String::from_utf8(bytes)?;
-        if is_valid_ledger_hash(&hash) {
-            return Ok(Self(hash));
-        }
-        bail!("Invalid ledger hash: {hash}")
-    }
-
-    pub fn from_bytes_or_panic(bytes: Vec<u8>) -> Self {
-        Self::from_bytes(bytes).expect("ledger hash bytes")
-    }
-}
-
-impl std::str::FromStr for LedgerHash {
-    type Err = anyhow::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if is_valid_ledger_hash(s) {
-            Ok(Self(s.to_string()))
-        } else {
-            bail!("Invalid ledger hash: {s}")
-        }
-    }
-}
-
-impl std::default::Default for LedgerHash {
-    fn default() -> Self {
-        Self("jxDEFAULTDEFAULTDEFAULTDEFAULTDEFAULTDEFAULTDEFAULT".into())
-    }
-}
-
-impl std::fmt::Display for LedgerHash {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
 }
 
 impl Ledger {
@@ -271,11 +220,6 @@ impl std::fmt::Debug for Ledger {
     }
 }
 
-pub fn is_valid_ledger_hash(input: &str) -> bool {
-    let prefix: String = input.chars().take(2).collect();
-    input.len() == LedgerHash::LEN && LedgerHash::PREFIX.contains(&prefix.as_str())
-}
-
 #[cfg(test)]
 mod tests {
     use super::{
@@ -284,7 +228,7 @@ mod tests {
             account::{AccountDiff, DelegationDiff, PaymentDiff, UpdateType},
             LedgerDiff,
         },
-        is_valid_ledger_hash,
+        hash::is_valid_ledger_hash,
         public_key::PublicKey,
         Ledger, LedgerHash,
     };
