+++ rust/src/unix_socket_server.rs
@@ -623,11 +623,12 @@ pub async fn handle_connection(
                             error!("Ledger at state hash {hash} is not in the store");
                             Some(format!("Ledger at state hash {hash} is not in the store"))
                         }
-                    } else if ledger::is_valid_ledger_hash(&hash) {
+                    } else if ledger::hash::is_valid_ledger_hash(&hash) {
                         trace!("{hash} is a ledger hash");
-                        if let Some(ledger) =
-                            db.get_staged_ledger_at_ledger_hash(&LedgerHash(hash.clone()), memoize)?
-                        {
+                        if let Some(ledger) = db.get_staged_ledger_at_ledger_hash(
+                            &LedgerHash::new_or_panic(hash.clone()),
+                            memoize,
+                        )? {
                             write_ledger(path, ledger, &hash)
                         } else {
                             error!("Ledger at ledger hash {hash} is not in the store");
@@ -675,7 +676,7 @@ pub async fn handle_connection(
             ClientCli::StakingLedgers(__) => match __ {
                 StakingLedgers::Hash { hash, path } => {
                     info!("Received staking-ledgers-hash command for {hash}");
-                    if ledger::is_valid_ledger_hash(&hash) {
+                    if ledger::hash::is_valid_ledger_hash(&hash) {
                         trace!("{hash} is a ledger hash");
                         if let Some(staking_ledger) =
                             db.get_staking_ledger(&hash.clone().into(), None, None)?
