+++ rust/tests/event/memoize_ledger.rs
@@ -44,7 +44,8 @@ async fn test() -> anyhow::Result<()> {
     // mainnet-7-3NLGcwFVQF1p1PrZpusw2fZwBe5HKXGtrGy1Vc4aPkeBtT8nMNUc.json
     let blockchain_length = 7;
     let state_hash = BlockHash("3NLGcwFVQF1p1PrZpusw2fZwBe5HKXGtrGy1Vc4aPkeBtT8nMNUc".into());
-    let ledger_hash = LedgerHash("jwFtwfnhd2PDb15c23uVgNqjS3PNVWP4HpZzYSVGQAv64Y2bdV5".into());
+    let ledger_hash =
+        LedgerHash::new_or_panic("jwFtwfnhd2PDb15c23uVgNqjS3PNVWP4HpZzYSVGQAv64Y2bdV5".into());
     assert!(indexer_store
         .get_staged_ledger_at_state_hash(&state_hash, true)?
         .is_some());
@@ -53,7 +54,8 @@ async fn test() -> anyhow::Result<()> {
     // mainnet-6-3NKqRR2BZFV7Ad5kxtGKNNL59neXohf4ZEC5EMKrrnijB1jy4R5v.json
     let blockchain_length_no = 6;
     let state_hash_no = BlockHash("3NKqRR2BZFV7Ad5kxtGKNNL59neXohf4ZEC5EMKrrnijB1jy4R5v".into());
-    let ledger_hash_no = LedgerHash("jxqrHaBcJzZAPW2rSa84chAxEHW7ot2GbqmRsWuNhwctZ8TFA2K".into());
+    let ledger_hash_no =
+        LedgerHash::new_or_panic("jxqrHaBcJzZAPW2rSa84chAxEHW7ot2GbqmRsWuNhwctZ8TFA2K".into());
     assert!(indexer_store
         .get_staged_ledger_at_state_hash(&state_hash_no, false)?
         .is_some());
@@ -80,7 +82,8 @@ async fn test() -> anyhow::Result<()> {
     // mainnet-4-3NL9qBsNibXPm5Nh8cSg5CCqrbzX5VUVY9gJzAbg7EVCF3hfhazG.json
     let blockchain_length = 4;
     let state_hash = BlockHash("3NL9qBsNibXPm5Nh8cSg5CCqrbzX5VUVY9gJzAbg7EVCF3hfhazG".into());
-    let ledger_hash = LedgerHash("jxw3wNhAUhyVT4AK4dGxtn4Kpx6pvk3AXVoi2A6BAEQweyV8Uwe".into());
+    let ledger_hash =
+        LedgerHash::new_or_panic("jxw3wNhAUhyVT4AK4dGxtn4Kpx6pvk3AXVoi2A6BAEQweyV8Uwe".into());
     assert!(indexer_store
         .get_staged_ledger_at_block_height(blockchain_length, true)?
         .is_some());
@@ -89,7 +92,8 @@ async fn test() -> anyhow::Result<()> {
     // mainnet-11-3NLMeYAFXxsmhSFtLHFxdtjGcfHTVFmBmBF8uTJvP4Ve5yEmxYeA.json
     let blockchain_length_no = 11;
     let state_hash_no = BlockHash("3NLMeYAFXxsmhSFtLHFxdtjGcfHTVFmBmBF8uTJvP4Ve5yEmxYeA".into());
-    let ledger_hash_no = LedgerHash("jxZVWjsyuQkPVSj7ZbqC8PPx8FXzHQjxUYA3bhvdnQQZ15jn7mR".into());
+    let ledger_hash_no =
+        LedgerHash::new_or_panic("jxZVWjsyuQkPVSj7ZbqC8PPx8FXzHQjxUYA3bhvdnQQZ15jn7mR".into());
     assert!(indexer_store
         .get_staged_ledger_at_block_height(blockchain_length_no, false)?
         .is_some());
